put event folders in components, fixed classNames on EventList and NewEvent modules, MVP event tickets complete
